backend/requirements.txt
Flask
flask-cors
flask-sqlalchemy
flask-migrate
pandas
pandas-ta
yfinance
vaderSentiment
newsapi-python
tensorflow
scikit-learn
joblib
python-dotenv
requests